Add software transition from launch script to orbit / deorbit
It'd be neat if the user could transition through a flight routine
without manually executing a series of scripts. For now I've built in a
terminal prompt to transition the user to the next applicable script.

In the future I could see this as part of a standalone controller which
runs the current launch automation.