Render positioned and scaled SOI and orbiting moon to match simulator
Using the OpenGL scale has allowed the rendered scene positions to
remain 1:1 with the simulator, but scaled in order to meet software
rendering constraints. A ratio of 0.00001:1.0 has been applied between
the visualizer and simulator respectively.

With these improvements I was able to render the SOI and a new orbiting
moon at their exact position and orbital apoapsis / periapsis. Work
might be done later to accurately emulate the body and orbit rotational
velocities.